create validation set from training set
Early stopping is used to prevent overfitting, but prior to this commit,
early stopping used the test set. This commit fixes this by creating a
validation set from the training set. Two new user parameters are added:
one to set the size of the training set and the other to set the seed
for random splitting. Sklearn is used for the splitting. We already
    imported it for something else, so it doesn't add any overhead to
    use it here.